Use Streams.
